fix(browser): split browser into 3 panels
